Extract #member from loop in #to_a
Refactoring.

Rename #member_row to #member
